feat!: make schema_version 64-bit
BREAKING CHANGE: previously the schema version was a 32-bit unsigned.
But it can become a problem for applications that widely create
vinyl spaces. As far as vinyl can't be set meta-temporary, its
creation always increments the schema. Let's make this counter 64-bit
unsigned to avoid the overflow problem. As a side effect be break
the IPROTO compatibility with the previous versions of Tarantool.

NO_DOC=core feature
NO_TEST=no Lua API

+++ src/box/lua/net_box.c
@@ -2491,9 +2491,9 @@ netbox_transport_do_auth(struct netbox_transport *transport,
  * On success invokes the 'did_fetch_schema' callback and returns the actual
  * schema version. On failure raises a Lua error.
  */
-static uint32_t
+static uint64_t
 netbox_transport_fetch_schema(struct netbox_transport *transport,
-			      struct lua_State *L, uint32_t schema_version)
+			      struct lua_State *L, uint64_t schema_version)
 {
 	if (!transport->opts.fetch_schema) {
 		return schema_version;
@@ -2614,9 +2614,9 @@ restart:
  * Returns the current schema version on schema change. On failure raises
  * a Lua error.
  */
-static uint32_t
+static uint64_t
 netbox_transport_process_requests(struct netbox_transport *transport,
-				  struct lua_State *L, uint32_t schema_version)
+				  struct lua_State *L, uint64_t schema_version)
 {
 	if (transport->state != NETBOX_ACTIVE &&
 	    transport->state != NETBOX_GRACEFUL_SHUTDOWN) {
@@ -2647,7 +2647,7 @@ netbox_connection_handler_f(struct lua_State *L)
 	struct netbox_transport *transport = (void *)lua_topointer(L, 1);
 	netbox_transport_do_id(transport, L);
 	netbox_transport_do_auth(transport, L);
-	uint32_t schema_version = 0;
+	uint64_t schema_version = 0;
 	while (true) {
 		schema_version = netbox_transport_fetch_schema(
 			transport, L, schema_version);
